% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drawDAG.R
\name{drawDAG}
\alias{drawDAG}
\title{drawDAG}
\usage{
drawDAG(dag, adj = NULL)
}
\arguments{
\item{dag}{A DAG string}

\item{adj}{nodes to adjust for}
}
\value{
a ggplot DAG
}
\description{
drawDAG
}
\examples{
g1 <- 'dag {
X [exposure, pos="0.000,0.000"]
Y [outcome, pos="2.000,0.000"]
Z [pos="1.000,1.000"]
Z -> X
Z -> Y
X -> Y
}'

drawDAG(g1)
drawDAG(g1, 'Z')
}
