% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/untidy_dagitty.R
\name{untidy_dagitty}
\alias{untidy_dagitty}
\title{untidy_dagitty}
\usage{
untidy_dagitty(td)
}
\arguments{
\item{td}{A \code{tidy_dagitty} object}
}
\value{
A named list containing two dagitty objects. The first element preserves the layout for plotting with \code{plot.dagitty()} or on \href{http://www.dagitty.net/dags.html}{dagitty.net}. The second element preserves the layout to plot with \code{ggdag} functions.
}
\description{
Revert a basic \code{tidy_dagitty} object to a \code{dagitty} object while preserving node positions
}
\examples{
library(dagitty)
library(ggdag)

dag <- dagitty( "dag {
  Y <- X <- Z1 <- V -> Z2 -> Y
  Z1 <- W1 <-> W2 -> Z2
  X <- W1 -> Y
  X <- W2 -> Y
  X [exposure]
  Y [outcome]
  }")
tidyDag <- ggdag::tidy_dagitty(dag)
ud <- untidy_dagitty(tidyDag)
plot(ud[["dagitty"]])
ggdag(ud[["ggdag"]])
}
