% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grader.R
\name{grader}
\alias{grader}
\title{Title}
\usage{
grader(submission, solution, nSol = 1)
}
\arguments{
\item{td}{}
}
\value{

}
\description{
Title
}
\examples{
sol <- list(c('Z1', 'Z2'), c('Z1', 'Z3'))
guess1 <- c('Z1')
guess2 <- c('Z1', 'Z2')
grader(guess1, sol, length(sol))
grader(guess2, sol, length(sol))
}
